requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-dateutil==2.8.2